Course duration added

+++ TrainingListCreateCourse.cpp
@@ -14,6 +14,8 @@ IMPLEMENT_DYNAMIC(TrainingListCreateCourse, CDialogEx)
 
 TrainingListCreateCourse::TrainingListCreateCourse(CWnd* pParent /*=nullptr*/)
 	: CDialogEx(IDD_TRAININGLIST_CREATE_COURSE, pParent)
+	, course_name_edit(_T(""))
+	, course_duration_edit(_T(""))
 {
 
 }
@@ -29,6 +31,8 @@ void TrainingListCreateCourse::DoDataExchange(CDataExchange* pDX)
 	DDX_Control(pDX, IDC_LIST3, received_skill_selected);
 	DDX_Control(pDX, IDC_LIST2, requirement_skill_selected);
 	DDX_Control(pDX, IDC_LIST5, skill_all);
+	DDX_Text(pDX, IDC_EDIT4, course_name_edit);
+	DDX_Text(pDX, IDC_EDIT3, course_duration_edit);
 }
 
 
@@ -38,6 +42,9 @@ BEGIN_MESSAGE_MAP(TrainingListCreateCourse, CDialogEx)
 	ON_NOTIFY(TCN_SELCHANGE, IDC_TAB1, &TrainingListCreateCourse::OnTcnSelchangeTab1)
 	ON_BN_CLICKED(IDC_BUTTON2, &TrainingListCreateCourse::OnBnClickedButton2)
 	ON_BN_CLICKED(IDC_BUTTON6, &TrainingListCreateCourse::OnBnClickedButton6)
+	ON_LBN_SELCHANGE(IDC_LIST2, &TrainingListCreateCourse::OnLbnSelchangeList2)
+	ON_LBN_SELCHANGE(IDC_LIST3, &TrainingListCreateCourse::OnLbnSelchangeList3)
+	ON_LBN_SELCHANGE(IDC_LIST5, &TrainingListCreateCourse::OnLbnSelchangeList5)
 END_MESSAGE_MAP()
 
 // Инициация диалога
@@ -49,6 +56,7 @@ BOOL TrainingListCreateCourse::OnInitDialog() {
 		FillAllSkills();
 		FillRequirementSkills();
 		FillReceivedSkills();
+		FillInputs();
 	} CATCH(CDBException, e) {
 		mainDlg.ReconnectDB();
 		TRY{
@@ -56,6 +64,7 @@ BOOL TrainingListCreateCourse::OnInitDialog() {
 			FillAllSkills();
 			FillRequirementSkills();
 			FillReceivedSkills();
+			FillInputs();
 		} CATCH(CDBException, e) {
 			AfxMessageBox(L"Database error: " + e->m_strError);
 		}
@@ -66,6 +75,53 @@ BOOL TrainingListCreateCourse::OnInitDialog() {
 	RedrawTab();
 
 	return TRUE;  // возврат значения TRUE, если фокус не передан элементу управления
+}
+
+// Инициация диалога
+void TrainingListCreateCourse::OnCancel() {
+	CDialogEx::OnCancel();
+	CString courseSqlString = L"", 
+			requirementSqlString = L"",
+			receivedSqlString = L"";
+	int courseDuration = 0;
+	UpdateData(true);
+
+	course_name_edit.Replace(L"\\", L"\\\\");
+	course_name_edit.Replace(L"\'", L"\\\'");
+	courseDuration = _ttoi(course_duration_edit);
+	if (courseDuration == 0) {
+		course_duration_edit = L"7";
+	}
+
+	if (_ttoi(course_id) > 0) {
+		courseSqlString = L"Update course set name = '" + course_name_edit + L"', duration = " + course_duration_edit + L" where course_id = " + course_id;
+	} else {
+		courseSqlString = L"Insert into course(name, duration) values ('" + course_name_edit + L"'," + course_duration_edit + L");";
+		requirementSqlString = L"Update requirement_skill set course_id = (select max(course_id) from course) where course_id = 0;";
+		receivedSqlString = L"Update received_skill set course_id = (select max(course_id) from course) where course_id = 0;";
+	}
+	
+	TRY{
+		database->ExecuteSQL(courseSqlString);
+		if (_ttoi(course_id) == 0) {
+			database->ExecuteSQL(requirementSqlString);
+			database->ExecuteSQL(receivedSqlString);
+		}
+	} CATCH(CDBException, e) {
+		CTrainingListDlg mainDlg;
+		mainDlg.ReconnectDB();
+		TRY{
+			database->ExecuteSQL(courseSqlString);
+			if (_ttoi(course_id) == 0) {
+				database->ExecuteSQL(requirementSqlString);
+				database->ExecuteSQL(receivedSqlString);
+			}
+		} CATCH(CDBException, e) {
+			AfxMessageBox(L"Database error: " + e->m_strError);
+		}
+		END_CATCH;
+	}
+	END_CATCH;
 }
 
 // Стилизация окна списка работников
@@ -85,7 +141,7 @@ void TrainingListCreateCourse::RedrawTab() {
 void TrainingListCreateCourse::FillSkillGroups(CString skill_group_id) {
 	CRecordset cr(database);
 	CString varVal, qStr;
-	int rowCount;
+	int rowCount, itemCount;
 
 	qStr = L"Select count(DISTINCT skill_group_id) cnt from skill where skill_id not in ( \
 					select skill_id from requirement_skill where course_id = " + course_id + L" \
@@ -107,6 +163,17 @@ void TrainingListCreateCourse::FillSkillGroups(CString skill_group_id) {
 	int i = 0;
 
 	skill_group.DeleteAllItems();
+	if (rowCount == 0)
+	{
+		skill_groups = new SkillGroup[1];
+		SkillGroup tmp{};
+		tmp.skill_group_id = L"0";
+		tmp.skill_group = L"";
+		skill_group.InsertItem(0, tmp.skill_group);
+		skill_groups[0] = tmp;
+		skill_group.SetCurSel(0);
+	}
+
 	while (!cr.IsEOF())
 	{
 		SkillGroup tmp{};
@@ -121,48 +188,44 @@ void TrainingListCreateCourse::FillSkillGroups(CString skill_group_id) {
 		i++;
 	}
 	cr.Close();
-	UpdateData(false);
 }
 
 // Заполнение окна умений
 void TrainingListCreateCourse::FillAllSkills() {
-	CRecordset cr(database);
-	CString varValue, qStr, firstQueryString, secondQueryString;
-	qStr = L"from skill where skill_group_id \
+	if (skill_group.GetCurSel() >= 0) {
+		CRecordset cr(database);
+		CString varValue, qStr;
+		qStr = L"from skill where skill_group_id \
 		= " + skill_groups[skill_group.GetCurSel()].skill_group_id + L" \
 		and skill_id not in ( \
 			select skill_id from requirement_skill where course_id = " + course_id + L" \
 			union \
 			select skill_id from received_skill where course_id = " + course_id + L") \
 		order by 1 desc";
-	firstQueryString = L"Select s.skill_id, s.name from ";
-	secondQueryString = L" rs \
-			left join skill s on rs.skill_id = s.skill_id \
-			where rs.course_id = " + course_id + L" \
-			order by 1 desc";
 
-	cr.Open(CRecordset::forwardOnly, L"Select count(*) cnt " + qStr);
-	cr.GetFieldValue(L"cnt", varValue);
-	int rowCount = _ttoi(varValue);
-	cr.Close();
-
-	skill_all.ResetContent();
-	if (rowCount > 0) {
-		cr.Open(CRecordset::forwardOnly, L"Select skill_id, name " + qStr);
-		skills_all = new Skill[rowCount];
-		int i = 0;
-
-		while (!cr.IsEOF())
-		{
-			Skill tmp{};
-			cr.GetFieldValue(L"skill_id", tmp.skill_id);
-			cr.GetFieldValue(L"name", tmp.skill);
-			cr.MoveNext();
-			skill_all.AddString(tmp.skill);
-			skills_all[i] = tmp;
-			i++;
-		}
+		cr.Open(CRecordset::forwardOnly, L"Select count(*) cnt " + qStr);
+		cr.GetFieldValue(L"cnt", varValue);
+		int rowCount = _ttoi(varValue);
 		cr.Close();
+
+		skill_all.ResetContent();
+		if (rowCount > 0) {
+			cr.Open(CRecordset::forwardOnly, L"Select skill_id, name " + qStr);
+			skills_all = new Skill[rowCount];
+			int i = 0;
+
+			while (!cr.IsEOF())
+			{
+				Skill tmp{};
+				cr.GetFieldValue(L"skill_id", tmp.skill_id);
+				cr.GetFieldValue(L"name", tmp.skill);
+				cr.MoveNext();
+				skill_all.AddString(tmp.skill);
+				skills_all[i] = tmp;
+				i++;
+			}
+			cr.Close();
+		}
 	}
 }
 
@@ -201,7 +264,6 @@ void TrainingListCreateCourse::FillRequirementSkills() {
 	}
 }
 
-
 // Заполнение окна умений
 void TrainingListCreateCourse::FillReceivedSkills() {
 	CRecordset cr(database);
@@ -237,6 +299,27 @@ void TrainingListCreateCourse::FillReceivedSkills() {
 	}
 }
 
+// Заполнение окна умений
+void TrainingListCreateCourse::FillInputs() {
+	if (_ttoi(course_id) > 0) {
+		CRecordset cr(database);
+		CString varValue;
+		
+		varValue = L"Select name, duration from course where course_id = " + course_id;
+		cr.Open(CRecordset::forwardOnly, varValue);
+
+		cr.GetFieldValue(L"name", course_name_edit);
+		cr.GetFieldValue(L"duration", course_duration_edit);
+
+		cr.Close();
+	}
+	else {
+		course_name_edit = L"";
+		course_duration_edit = L"7";
+	}
+	UpdateData(false);
+}
+
 // Функция для передачи в этот класс БД
 void TrainingListCreateCourse::SetDB(CDatabase* database) {
 	this->database = database;
@@ -245,57 +328,76 @@ void TrainingListCreateCourse::SetDB(CDatabase* database) {
 // Добавить скилл в требуемые
 void TrainingListCreateCourse::OnBnClickedButton1()
 {
-	CString SqlString, err = L"";
-	CString skill_id = skills_all[skill_all.GetCurSel()].skill_id;
-	SqlString = L"INSERT INTO requirement_skill (course_id, skill_id) VALUES (" + course_id + L"," + skill_id + L")";
+	if (skill_all.GetCurSel() >= 0) {
+		CString SqlString, skill_group_id, err = L"";
+		CString skill_id = skills_all[skill_all.GetCurSel()].skill_id;
+		SqlString = L"INSERT INTO requirement_skill (course_id, skill_id) VALUES (" + course_id + L"," + skill_id + L")";
 
-	TRY{
-		database->ExecuteSQL(SqlString);
-	} CATCH(CDBException, e) {
-		CTrainingListDlg mainDlg;
-		mainDlg.ReconnectDB();
 		TRY{
 			database->ExecuteSQL(SqlString);
 		} CATCH(CDBException, e) {
-			err = e->m_strError;
+			CTrainingListDlg mainDlg;
+			mainDlg.ReconnectDB();
+			TRY{
+				database->ExecuteSQL(SqlString);
+			} CATCH(CDBException, e) {
+				err = e->m_strError;
+			}
+			END_CATCH;
 		}
 		END_CATCH;
-	}
-	END_CATCH;
 
-	if (err == "") {
-		FillRequirementSkills();
-		FillSkillGroups(skill_groups[skill_group.GetCurSel()].skill_group_id);
-		FillAllSkills();
+		if (err == "") {
+			skill_group_id = L"";
+			if (skill_group.GetCurSel() >= 0) {
+				skill_group_id = skill_groups[skill_group.GetCurSel()].skill_group_id;
+			}
+			FillRequirementSkills();
+			FillSkillGroups(skill_group_id);
+			FillAllSkills();
+		}
 	}
+	GetDlgItem(IDC_BUTTON1)->EnableWindow(false);
+	GetDlgItem(IDC_BUTTON2)->EnableWindow(false);
+	GetDlgItem(IDC_BUTTON5)->EnableWindow(false);
 }
 
 // Убрать скилл из требуемых
 void TrainingListCreateCourse::OnBnClickedButton5()
 {
-	CString SqlString, err = L"";
-	CString skill_id = requirement_skills_selected[requirement_skill_selected.GetCurSel()].skill_id;
-	SqlString = L"Delete from requirement_skill where course_id = " + course_id + L" and skill_id = " + skill_id;
+	if (requirement_skill_selected.GetCurSel() >= 0) {
+		CString SqlString, skill_group_id, err = L"";
+		CString skill_id = requirement_skills_selected[requirement_skill_selected.GetCurSel()].skill_id;
+		SqlString = L"Delete from requirement_skill where course_id = " + course_id + L" and skill_id = " + skill_id;
 
-	TRY{
-		database->ExecuteSQL(SqlString);
-	} CATCH(CDBException, e) {
-		CTrainingListDlg mainDlg;
-		mainDlg.ReconnectDB();
 		TRY{
 			database->ExecuteSQL(SqlString);
 		} CATCH(CDBException, e) {
-			err = e->m_strError;
+			CTrainingListDlg mainDlg;
+			mainDlg.ReconnectDB();
+			TRY{
+				database->ExecuteSQL(SqlString);
+			} CATCH(CDBException, e) {
+				err = e->m_strError;
+			}
+			END_CATCH;
 		}
 		END_CATCH;
-	}
-	END_CATCH;
 
-	if (err == "") {
-		FillSkillGroups(skill_groups[skill_group.GetCurSel()].skill_group_id);
-		FillRequirementSkills();
-		FillAllSkills();
+		if (err == "") {
+			skill_group_id = L"";
+			if (skill_group.GetCurSel() >= 0) {
+				skill_group_id = skill_groups[skill_group.GetCurSel()].skill_group_id;
+			}
+			FillSkillGroups(skill_group_id);
+			FillRequirementSkills();
+			FillAllSkills();
+		}
 	}
+
+	GetDlgItem(IDC_BUTTON1)->EnableWindow(false);
+	GetDlgItem(IDC_BUTTON2)->EnableWindow(false);
+	GetDlgItem(IDC_BUTTON5)->EnableWindow(false);
 }
 
 // Смена скилл группы
@@ -314,6 +416,8 @@ void TrainingListCreateCourse::OnTcnSelchangeTab1(NMHDR* pNMHDR, LRESULT* pResul
 		END_CATCH;
 	}
 	END_CATCH;
+	GetDlgItem(IDC_BUTTON1)->EnableWindow(false);
+	GetDlgItem(IDC_BUTTON2)->EnableWindow(false);
 	
 	*pResult = 0;
 }
@@ -321,55 +425,94 @@ void TrainingListCreateCourse::OnTcnSelchangeTab1(NMHDR* pNMHDR, LRESULT* pResul
 // Добавить скилл в получаемые
 void TrainingListCreateCourse::OnBnClickedButton2()
 {
-	CString SqlString, err = L"";
-	CString skill_id = skills_all[skill_all.GetCurSel()].skill_id;
-	SqlString = L"INSERT INTO received_skill (course_id, skill_id) VALUES (" + course_id + L"," + skill_id + L")";
+	if (skill_all.GetCurSel() >= 0) {
+		CString SqlString, skill_group_id, err = L"";
+		CString skill_id = skills_all[skill_all.GetCurSel()].skill_id;
+		SqlString = L"INSERT INTO received_skill (course_id, skill_id) VALUES (" + course_id + L"," + skill_id + L")";
 
-	TRY{
-		database->ExecuteSQL(SqlString);
-	} CATCH(CDBException, e) {
-		CTrainingListDlg mainDlg;
-		mainDlg.ReconnectDB();
 		TRY{
 			database->ExecuteSQL(SqlString);
 		} CATCH(CDBException, e) {
-			err = e->m_strError;
+			CTrainingListDlg mainDlg;
+			mainDlg.ReconnectDB();
+			TRY{
+				database->ExecuteSQL(SqlString);
+			} CATCH(CDBException, e) {
+				err = e->m_strError;
+			}
+			END_CATCH;
 		}
 		END_CATCH;
-	}
-	END_CATCH;
 
-	if (err == "") {
-		FillReceivedSkills();
-		FillSkillGroups(skill_groups[skill_group.GetCurSel()].skill_group_id);
-		FillAllSkills();
+		if (err == "") {
+			skill_group_id = L"";
+			if (skill_group.GetCurSel() >= 0) {
+				skill_group_id = skill_groups[skill_group.GetCurSel()].skill_group_id;
+			}
+			FillReceivedSkills();
+			FillSkillGroups(skill_group_id);
+			FillAllSkills();
+		}
 	}
+
+	GetDlgItem(IDC_BUTTON1)->EnableWindow(false);
+	GetDlgItem(IDC_BUTTON2)->EnableWindow(false);
+	GetDlgItem(IDC_BUTTON6)->EnableWindow(false);
 }
 
 // Убрать скилл из получаемых
 void TrainingListCreateCourse::OnBnClickedButton6()
 {
-	CString SqlString, err = L"";
-	CString skill_id = received_skills_selected[received_skill_selected.GetCurSel()].skill_id;
-	SqlString = L"Delete from received_skill where course_id = " + course_id + L" and skill_id = " + skill_id;
+	if (received_skill_selected.GetCurSel() >= 0) {
+		CString SqlString, skill_group_id, err = L"";
+		CString skill_id = received_skills_selected[received_skill_selected.GetCurSel()].skill_id;
+		SqlString = L"Delete from received_skill where course_id = " + course_id + L" and skill_id = " + skill_id;
 
-	TRY{
-		database->ExecuteSQL(SqlString);
-	} CATCH(CDBException, e) {
-		CTrainingListDlg mainDlg;
-		mainDlg.ReconnectDB();
 		TRY{
 			database->ExecuteSQL(SqlString);
 		} CATCH(CDBException, e) {
-			err = e->m_strError;
+			CTrainingListDlg mainDlg;
+			mainDlg.ReconnectDB();
+			TRY{
+				database->ExecuteSQL(SqlString);
+			} CATCH(CDBException, e) {
+				err = e->m_strError;
+			}
+			END_CATCH;
 		}
 		END_CATCH;
-	}
-	END_CATCH;
 
-	if (err == "") {
-		FillSkillGroups(skill_groups[skill_group.GetCurSel()].skill_group_id);
-		FillReceivedSkills();
-		FillAllSkills();
+		if (err == "") {
+			skill_group_id = L"";
+			if (skill_group.GetCurSel() >= 0) {
+				skill_group_id = skill_groups[skill_group.GetCurSel()].skill_group_id;
+			}
+			FillSkillGroups(skill_group_id);
+			FillReceivedSkills();
+			FillAllSkills();
+		}
 	}
+
+	GetDlgItem(IDC_BUTTON1)->EnableWindow(false);
+	GetDlgItem(IDC_BUTTON2)->EnableWindow(false);
+	GetDlgItem(IDC_BUTTON6)->EnableWindow(false);
+}
+
+
+void TrainingListCreateCourse::OnLbnSelchangeList2()
+{
+	GetDlgItem(IDC_BUTTON5)->EnableWindow(requirement_skill_selected.GetCurSel() >= 0);
+}
+
+
+void TrainingListCreateCourse::OnLbnSelchangeList3()
+{
+	GetDlgItem(IDC_BUTTON6)->EnableWindow(received_skill_selected.GetCurSel() >= 0);
+}
+
+
+void TrainingListCreateCourse::OnLbnSelchangeList5()
+{
+	GetDlgItem(IDC_BUTTON1)->EnableWindow(skill_all.GetCurSel() >= 0);
+	GetDlgItem(IDC_BUTTON2)->EnableWindow(skill_all.GetCurSel() >= 0);
 }
